requirements.txt
flask
flask-cors
gunicorn
playwright
google-generativeai